requirements.txt
cron-descriptor==1.2.24
croniter==0.3.31
Django==3.0.7
django-compressor==2.4
psycopg2==2.8.5
pytz==2020.1
requests==2.23.0
statsd==3.3.0
selenium==3.141.0
pytest==6.0.1
allure-pytest==2.8.11
allure-python-commons==2.8.11
flake8==3.8.3
